Feat: Повторяющиеся функции вынесены в отдельный компонент
